refactor(replication): isolate ecstore boundary imports
