+++ crates/ecstore/src/bucket/replication/replication_pool.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::datatypes::ResyncStatusType;
 use super::replication_config_store::ReplicationConfigStore;
 use super::replication_error_boundary::Error as EcstoreError;
 use super::replication_filemeta_boundary::{
@@ -30,6 +29,7 @@ use super::replication_queue_boundary::{
     initial_worker_counts, large_worker_backpressure_resize, mrf_worker_size_to_count, replication_backpressure_recommendation,
     replication_heal_queue_action, resized_worker_counts, should_queue_large_object, worker_queue_for_replication_type,
 };
+use super::replication_resync_boundary::ResyncStatusType;
 use super::replication_resync_boundary::{
     BucketReplicationResyncStatus, ResyncOpts, TargetReplicationResyncStatus, decode_mrf_file, decode_resync_file,
     encode_mrf_file, should_auto_resume_resync,
