+++ crates/ecstore/src/bucket/replication/replication_resyncer.rs
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::config::{ObjectOpts, ReplicationConfigurationExt as _};
-use super::datatypes::ResyncStatusType;
 use super::replication_bandwidth_boundary;
+use super::replication_config_boundary::{ObjectOpts, ReplicationConfigurationExt as _};
 use super::replication_config_store::ReplicationConfigStore;
 use super::replication_error_boundary::{Result, is_err_object_not_found, is_err_version_not_found};
 use super::replication_event_sink::{EventArgs, send_event, send_local_event};
@@ -35,6 +34,7 @@ use super::replication_object_decision_boundary::{
     replication_multipart_complete_actual_size, replication_multipart_part_plan, should_retry_delete_marker_purge,
 };
 use super::replication_queue_boundary::DeletedObjectReplicationInfo;
+use super::replication_resync_boundary::ResyncStatusType;
 use super::replication_resync_boundary::{
     BucketReplicationResyncStatus, ResyncOpts, TargetReplicationResyncStatus, encode_resync_file, is_version_id_mismatch,
     resync_state_accepts_update, should_count_head_proxy_failure,
